Gotovo

+++ vite-project/src/pages/Products.jsx
@@ -1,11 +1,12 @@
-import React from 'react'
+import React, { useState } from 'react'
 import useFetch from '../hooks/useFetch'
 import {deleteProduct, productsService} from '../services/productsService'
 import ProductTable from '../components/ProductTable';
+import { Link } from 'react-router-dom';
 
 const Products = () => {
   const {products,error,loading,refetch} = useFetch(productsService);  
-  
+  const [search, setSearch] = useState("");
   if(loading) {
     return <p>Ucitavanje podataka</p>
   }
@@ -21,6 +22,7 @@ const Products = () => {
 
   const handleDelete = async(id) => {
     try {
+      if (!window.confirm("Sigurno?")) return;
       await deleteProduct(id);
       refetch();
     } catch (error) {
@@ -28,9 +30,20 @@ const Products = () => {
     }
   }
 
+  const handleClick = (e) => {
+    setSearch(e.target.value);
+  }
+
+  const filtered=products.filter(p=>p.name.toLowerCase().includes(search.toLowerCase()))
+
+
   return (
     <div>
-      <ProductTable products={products} onDelete={handleDelete}/>
+      <Link to={"/products/new"}>Create new product</Link>
+      <p>Pretrazi:</p>
+      <input value={search} onChange={handleClick}></input>
+      
+      <ProductTable products={filtered} onDelete={handleDelete}/>
     </div>
   )
 }
